rename app to server
